+++ pennylane_lightning/core/src/simulators/lightning_tensor/tncuda/MPSTNCuda.hpp
@@ -29,6 +29,7 @@
 
 #include "DataBuffer.hpp"
 #include "DevTag.hpp"
+#include "MPOTNCuda.hpp"
 #include "TNCudaBase.hpp"
 #include "TensorCuda.hpp"
 #include "TensornetBase.hpp"
@@ -72,6 +73,9 @@ class MPSTNCuda final : public TNCudaBase<Precision, MPSTNCuda<Precision>> {
     std::vector<TensorCuda<Precision>> tensors_;
 
     std::vector<TensorCuda<Precision>> tensors_out_;
+
+    std::vector<std::shared_ptr<MPOTNCuda<Precision>>> mpos_;
+    std::vector<std::size_t> mpo_ids_;
 
   public:
     using CFP_t = decltype(cuUtil::getCudaType(Precision{}));
@@ -233,6 +237,77 @@ class MPSTNCuda final : public TNCudaBase<Precision, MPSTNCuda<Precision>> {
     };
 
     /**
+     * @brief Apply an MPO operator with the gate's MPO decomposition data
+     * provided by the user to the compute graph.
+     *
+     * This API only works for the MPS backend.
+     *
+     * @param tensors The MPO representation of a gate. Each element in the
+     * outer vector represents a MPO tensor site.
+     * @param wires The wire indices of the gate acts on. The size of this
+     * vector should match the size of the `tensors` vector.
+     * @param max_mpo_bond_dim The maximum bond dimension of the MPO operator.
+     */
+    void applyMPOOperation(const std::vector<std::vector<ComplexT>> &tensors,
+                           const std::vector<std::size_t> &wires,
+                           const std::size_t max_mpo_bond_dim) {
+        PL_ABORT_IF_NOT(
+            tensors.size() == wires.size(),
+            "The number of tensors should be equal to the number of "
+            "wires.");
+
+        // Create a queue of wire pairs to apply SWAP gates and MPO local target
+        // wires
+        const auto [local_wires, swap_wires_queue] =
+            create_swap_wire_pair_queue(wires);
+
+        // Apply SWAP gates to ensure the following MPO operator targeting at
+        // local wires
+        if (swap_wires_queue.size() > 0) {
+            for_each(swap_wires_queue.begin(), swap_wires_queue.end(),
+                     [this](const auto &swap_wires) {
+                         for_each(swap_wires.begin(), swap_wires.end(),
+                                  [this](const auto &wire_pair) {
+                                      BaseType::applyOperation(
+                                          "SWAP", wire_pair, false);
+                                  });
+                     });
+        }
+
+        // Create a MPO object based on the host data from the user
+        mpos_.emplace_back(std::make_shared<MPOTNCuda<Precision>>(
+            tensors, local_wires, max_mpo_bond_dim, BaseType::getNumQubits(),
+            BaseType::getTNCudaHandle(), BaseType::getCudaDataType(),
+            BaseType::getDevTag()));
+
+        // Append the MPO operator to the compute graph
+        // Note MPO operator only works for local target wires as of v24.08
+        int64_t operatorId;
+        PL_CUTENSORNET_IS_SUCCESS(cutensornetStateApplyNetworkOperator(
+            /* const cutensornetHandle_t */ BaseType::getTNCudaHandle(),
+            /* cutensornetState_t */ BaseType::getQuantumState(),
+            /* cutensornetNetworkOperator_t */ mpos_.back()->getMPOOperator(),
+            /* const int32_t immutable */ 1,
+            /* const int32_t adjoint */ 0,
+            /* const int32_t unitary */ 1,
+            /* int64_t * operatorId*/ &operatorId));
+
+        mpo_ids_.push_back(static_cast<std::size_t>(operatorId));
+
+        // Apply SWAP gates to restore the original wire order
+        if (swap_wires_queue.size() > 0) {
+            for_each(swap_wires_queue.rbegin(), swap_wires_queue.rend(),
+                     [this](const auto &swap_wires) {
+                         for_each(swap_wires.rbegin(), swap_wires.rend(),
+                                  [this](const auto &wire_pair) {
+                                      BaseType::applyOperation(
+                                          "SWAP", wire_pair, false);
+                                  });
+                     });
+        }
+    }
+
+    /**
      * @brief Append MPS final state to the quantum circuit.
      *
      * @param cutoff Cutoff value for SVD decomposition. Default is 0.
@@ -275,6 +350,21 @@ class MPSTNCuda final : public TNCudaBase<Precision, MPSTNCuda<Precision>> {
             /* cutensornetStateAttributes_t */ svd_cutoff_mode,
             /* const void * */ &cutoff,
             /* std::size_t */ sizeof(cutoff)));
+
+        // MPO configurations
+        // Note that CUTENSORNET_STATE_MPO_APPLICATION_INEXACT is applied if the
+        // `cutoff` value is not set to 0 for the MPO application.
+        cutensornetStateMPOApplication_t mpo_attribute =
+            (cutoff == 0) ? CUTENSORNET_STATE_MPO_APPLICATION_EXACT
+                          : CUTENSORNET_STATE_MPO_APPLICATION_INEXACT;
+
+        PL_CUTENSORNET_IS_SUCCESS(cutensornetStateConfigure(
+            /* const cutensornetHandle_t */ BaseType::getTNCudaHandle(),
+            /* cutensornetState_t */ BaseType::getQuantumState(),
+            /* cutensornetStateAttributes_t */
+            CUTENSORNET_STATE_CONFIG_MPS_MPO_APPLICATION,
+            /* const void * */ &mpo_attribute,
+            /* std::size_t */ sizeof(mpo_attribute)));
 
         BaseType::computeState(
             const_cast<int64_t **>(getSitesExtentsPtr().data()),